image-captioning/requirements.txt
gradio>=4.44.0
transformers>=4.30.0
torch>=1.13.0
numpy>=1.24.0
Pillow>=9.5.0